casts

+++ week1/blocked.cpp
@@ -24,7 +24,7 @@ void build(int* arr, int pos, int* numbers, int start, int end)
     curleaf = d * parentleaf + i;
     // printf("cur leaf = %d\n", curleaf);
 
-    int index = start + ((double)(end - start + 1) / d) * i;
+    int index = (int)(start + ((double)(end - start + 1) / d) * i);
 
     // printf("index = %d\n", index);
     arr[B * parentleaf + (i - 1)] = numbers[index];
@@ -167,7 +167,7 @@ int bs_bs_search_iter(int q, int* arr, int n, int block)
 int main(int argc, char* argv[])
 {
   int i = 5;
-  int n = pow(d, i) - 1;
+  int n = (int) pow((double) d, i) - 1;
   int* arr = (int*) malloc(n * sizeof(int));
   int* numbers = (int*) malloc(n * sizeof(int));
